backend/requirements.txt
pandas
beautifulsoup4==4.8.0
bs4==0.0.1
decorator==4.3.2
defusedxml==0.5.0
Deprecated==1.2.7
es-core-news-md==2.1.0
es-core-news-sm==2.1.0
falcon==2.0.0
fastrlock==0.4
favicon==0.7.0
fbmessenger==6.0.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SocketIO==4.1.0
future==0.17.1
h5py==2.9.0
html5lib==1.0.1
httplib2==0.9.2
httptools==0.0.13
Jinja2==2.10.1
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
lxml==4.4.1
murmurhash==1.0.2
pdfminer3==2018.12.3.0
pdfminer3k==1.3.1
pdfrw==0.4
pyparsing==2.3.1
PyPDF2==1.26.0
pytest==5.1.2
python-docx==0.8.10
spacy==2.1.8
SQLAlchemy==1.3.8
sqlparse==0.3.0
srsly==0.1.0
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.3
wadllib==1.3.2
wasabi==0.2.2
webexteamssdk==1.2
websocket-client==0.54.0
websockets==6.0
widgetsnbextension==3.4.2
wrapt==1.11.1
xkit==0.0.0
xlrd==1.2.0
xlwt==1.3.0
